refactor: split ConversationViewModelTest tests
