+++ app/src/test/kotlin/com/wire/android/ui/home/conversations/ConversationsViewModelArrangement.kt
@@ -1,6 +1,5 @@
 package com.wire.android.ui.home.conversations
 
-import android.content.res.Resources
 import androidx.lifecycle.SavedStateHandle
 import com.wire.android.config.TestDispatcherProvider
 import com.wire.android.config.mockUri
@@ -12,8 +11,6 @@ import com.wire.android.ui.home.conversations.model.MessageSource
 import com.wire.android.ui.home.conversations.model.MessageStatus
 import com.wire.android.ui.home.conversations.model.MessageTime
 import com.wire.android.ui.home.conversations.model.UIMessage
-import com.wire.android.ui.home.conversations.usecase.GetMessagesForConversationUseCase
-import com.wire.android.util.FileManager
 import com.wire.android.util.ui.UIText
 import com.wire.android.util.ui.WireSessionImageLoader
 import com.wire.kalium.logic.CoreFailure
@@ -32,17 +29,13 @@ import com.wire.kalium.logic.data.user.UserAssetId
 import com.wire.kalium.logic.data.user.UserAvailabilityStatus
 import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.data.user.type.UserType
-import com.wire.kalium.logic.feature.asset.GetMessageAssetUseCase
 import com.wire.kalium.logic.feature.asset.SendAssetMessageResult
 import com.wire.kalium.logic.feature.asset.SendAssetMessageUseCase
-import com.wire.kalium.logic.feature.asset.UpdateAssetMessageDownloadStatusUseCase
-import com.wire.kalium.logic.feature.call.AnswerCallUseCase
 import com.wire.kalium.logic.feature.call.usecase.EndCallUseCase
 import com.wire.kalium.logic.feature.call.usecase.ObserveEstablishedCallsUseCase
 import com.wire.kalium.logic.feature.call.usecase.ObserveOngoingCallsUseCase
 import com.wire.kalium.logic.feature.conversation.GetSecurityClassificationTypeUseCase
 import com.wire.kalium.logic.feature.conversation.IsSelfUserMemberResult
-import com.wire.kalium.logic.feature.conversation.ObserveConversationDetailsUseCase
 import com.wire.kalium.logic.feature.conversation.ObserveIsSelfUserMemberUseCase
 import com.wire.kalium.logic.feature.conversation.UpdateConversationReadDateUseCase
 import com.wire.kalium.logic.feature.message.DeleteMessageUseCase
@@ -56,11 +49,8 @@ import io.mockk.coEvery
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.mockk
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.consumeAsFlow
 import kotlinx.coroutines.flow.emptyFlow
 import kotlinx.coroutines.flow.flowOf
-import kotlinx.coroutines.flow.map
 import okio.Path
 import okio.buffer
 
@@ -73,12 +63,9 @@ internal class ConversationsViewModelArrangement {
         MockKAnnotations.init(this, relaxUnitFun = true)
         mockUri()
         every { savedStateHandle.get<String>(any()) } returns conversationId.toString()
-        every { savedStateHandle.set(any(), any<String>()) } returns Unit
 
         // Default empty values
-        coEvery { observeConversationDetails(any()) } returns flowOf()
         coEvery { getSelfUserTeam() } returns flowOf()
-        coEvery { getMessagesForConversationUseCase(any()) } returns flowOf(listOf())
         every { isFileSharingEnabledUseCase() } returns FileSharingStatus(null, null)
         coEvery { observeOngoingCallsUseCase() } returns flowOf(listOf())
         coEvery { observeEstablishedCallsUseCase() } returns flowOf(listOf())
@@ -104,40 +91,16 @@ internal class ConversationsViewModelArrangement {
     lateinit var sendAssetMessage: SendAssetMessageUseCase
 
     @MockK
-    lateinit var getMessageAsset: GetMessageAssetUseCase
-
-    @MockK
     lateinit var deleteMessage: DeleteMessageUseCase
-
-    @MockK
-    lateinit var observeConversationDetails: ObserveConversationDetailsUseCase
-
-    @MockK
-    lateinit var updateAssetMessageDownloadStatus: UpdateAssetMessageDownloadStatusUseCase
 
     @MockK
     lateinit var getSelfUserTeam: GetSelfTeamUseCase
 
     @MockK
-    lateinit var fileManager: FileManager
-
-    @MockK
-    lateinit var getMessagesForConversationUseCase: GetMessagesForConversationUseCase
-
-    @MockK
     lateinit var isFileSharingEnabledUseCase: IsFileSharingEnabledUseCase
 
     @MockK
-    lateinit var resources: Resources
-
-    @MockK
-    lateinit var uiText: UIText
-
-    @MockK
     lateinit var observeOngoingCallsUseCase: ObserveOngoingCallsUseCase
-
-    @MockK
-    lateinit var answerCallUseCase: AnswerCallUseCase
 
     @MockK
     private lateinit var wireSessionImageLoader: WireSessionImageLoader
@@ -162,110 +125,48 @@ internal class ConversationsViewModelArrangement {
 
     private val fakeKaliumFileSystem = FakeKaliumFileSystem()
 
-    private val conversationDetailsChannel = Channel<ConversationDetails>(capacity = Channel.UNLIMITED)
-
-    private val messagesChannel = Channel<List<UIMessage>>(capacity = Channel.UNLIMITED)
-
     private val viewModel by lazy {
         MessageComposerViewModel(
             savedStateHandle = savedStateHandle,
             navigationManager = navigationManager,
             qualifiedIdMapper = qualifiedIdMapper,
-            observeConversationDetails = observeConversationDetails,
             sendTextMessage = sendTextMessage,
             sendAssetMessage = sendAssetMessage,
-            getMessageAsset = getMessageAsset,
             deleteMessage = deleteMessage,
             dispatchers = TestDispatcherProvider(),
-            updateAssetMessageDownloadStatus = updateAssetMessageDownloadStatus,
             getSelfUserTeam = getSelfUserTeam,
-            fileManager = fileManager,
-            getMessageForConversation = getMessagesForConversationUseCase,
             isFileSharingEnabled = isFileSharingEnabledUseCase,
-            observeOngoingCalls = observeOngoingCallsUseCase,
-            answerCall = answerCallUseCase,
             wireSessionImageLoader = wireSessionImageLoader,
             kaliumFileSystem = fakeKaliumFileSystem,
-            observeEstablishedCalls = observeEstablishedCallsUseCase,
-            endCall = endCall,
             updateConversationReadDateUseCase = updateConversationReadDateUseCase,
             observeIsSelfConversationMember = observeIsSelfUserMemberUseCase,
-            observeSyncState = observeSyncState,
             getConversationClassifiedType = getSecurityClassificationType
         )
     }
 
-    suspend fun withSuccessfulViewModelInit(): ConversationsViewModelArrangement {
+    suspend fun withSuccessfulViewModelInit() = apply {
         coEvery { isFileSharingEnabledUseCase() } returns FileSharingStatus(null, null)
-        coEvery { getMessagesForConversationUseCase(any()) } returns messagesChannel.consumeAsFlow()
         coEvery { observeOngoingCallsUseCase() } returns emptyFlow()
         coEvery { observeEstablishedCallsUseCase() } returns emptyFlow()
         coEvery { observeIsSelfUserMemberUseCase(any()) } returns flowOf(IsSelfUserMemberResult.Success(true))
-        return this
     }
 
-    fun withStoredAsset(dataPath: Path, dataContent: ByteArray): ConversationsViewModelArrangement {
+    fun withStoredAsset(dataPath: Path, dataContent: ByteArray) = apply {
         fakeKaliumFileSystem.sink(dataPath).buffer().use {
             it.write(dataContent)
         }
-        return this
     }
 
-    suspend fun withMessagesUpdate(messages: List<UIMessage>): ConversationsViewModelArrangement {
-        coEvery { getMessagesForConversationUseCase(any()) } returns messagesChannel.consumeAsFlow()
-        messagesChannel.send(messages)
-
-        return this
-    }
-
-    suspend fun withConversationDetailUpdate(conversationDetails: ConversationDetails): ConversationsViewModelArrangement {
-        coEvery { observeConversationDetails(any()) } returns conversationDetailsChannel.consumeAsFlow().map {
-            ObserveConversationDetailsUseCase.Result.Success(it)
-        }
-        conversationDetailsChannel.send(conversationDetails)
-        coEvery {
-            qualifiedIdMapper.fromStringToQualifiedID("id@domain")
-        } returns QualifiedID("id", "domain")
-        return this
-    }
-
-    fun withSuccessfulSendAttachmentMessage(): ConversationsViewModelArrangement {
+    fun withSuccessfulSendAttachmentMessage() = apply {
         coEvery { sendAssetMessage(any(), any(), any(), any(), any(), any(), any()) } returns SendAssetMessageResult.Success
-        return this
     }
 
-    fun withFailureOnDeletingMessages(): ConversationsViewModelArrangement {
+    fun withFailureOnDeletingMessages() = apply {
         coEvery { deleteMessage(any(), any(), any()) } returns Either.Left(CoreFailure.Unknown(null))
         return this
     }
 
-    fun withSuccessfulSaveAssetMessage(
-        assetName: String,
-        assetDataPath: Path,
-        assetSize: Long,
-        messageId: String
-    ): ConversationsViewModelArrangement {
-        viewModel.showOnAssetDownloadedDialog(assetName, assetDataPath, assetSize, messageId)
-        coEvery { fileManager.saveToExternalStorage(any(), any(), any(), any()) }.answers {
-            viewModel.hideOnAssetDownloadedDialog()
-        }
-        return this
-    }
-
-    fun withSuccessfulOpenAssetMessage(
-        assetName: String,
-        assetDataPath: Path,
-        assetSize: Long,
-        messageId: String
-    ): ConversationsViewModelArrangement {
-        viewModel.showOnAssetDownloadedDialog(assetName, assetDataPath, assetSize, messageId)
-        every { fileManager.openWithExternalApp(any(), any(), any()) }.answers {
-            viewModel.hideOnAssetDownloadedDialog()
-        }
-        return this
-    }
-
-    fun withTeamUser(userTeam: Team): ConversationsViewModelArrangement {
+    fun withTeamUser(userTeam: Team) = apply {
         coEvery { getSelfUserTeam() } returns flowOf(userTeam)
         return this
     }
