+++ app/src/test/kotlin/com/wire/android/ui/home/conversations/info/ConversationInfoViewModelArrangement.kt
@@ -1,0 +1,72 @@
+package com.wire.android.ui.home.conversations.info
+
+import androidx.lifecycle.SavedStateHandle
+import com.wire.android.config.mockUri
+import com.wire.android.navigation.NavigationManager
+import com.wire.android.util.ui.WireSessionImageLoader
+import com.wire.kalium.logic.data.conversation.ConversationDetails
+import com.wire.kalium.logic.data.id.ConversationId
+import com.wire.kalium.logic.data.id.QualifiedID
+import com.wire.kalium.logic.data.id.QualifiedIdMapper
+import com.wire.kalium.logic.feature.conversation.ObserveConversationDetailsUseCase
+import io.mockk.MockKAnnotations
+import io.mockk.coEvery
+import io.mockk.every
+import io.mockk.impl.annotations.MockK
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.consumeAsFlow
+import kotlinx.coroutines.flow.map
+
+
+class ConversationInfoViewModelArrangement {
+
+    val conversationId: ConversationId = ConversationId("some-dummy-value", "some.dummy.domain")
+
+    private val conversationDetailsChannel = Channel<ConversationDetails>(capacity = Channel.UNLIMITED)
+
+    @MockK
+    lateinit var qualifiedIdMapper: QualifiedIdMapper
+
+    @MockK
+    private lateinit var savedStateHandle: SavedStateHandle
+
+    @MockK
+    lateinit var navigationManager: NavigationManager
+
+    @MockK
+    lateinit var observeConversationDetails: ObserveConversationDetailsUseCase
+
+    @MockK
+    private lateinit var wireSessionImageLoader: WireSessionImageLoader
+
+    private val viewModel: ConversationInfoViewModel by lazy {
+        ConversationInfoViewModel(
+            qualifiedIdMapper,
+            savedStateHandle,
+            navigationManager,
+            observeConversationDetails,
+            wireSessionImageLoader
+        )
+    }
+
+    init {
+        MockKAnnotations.init(this, relaxUnitFun = true)
+        mockUri()
+        every { savedStateHandle.get<String>(any()) } returns conversationId.toString()
+        every {
+            qualifiedIdMapper.fromStringToQualifiedID("some-dummy-value@some.dummy.domain")
+        } returns QualifiedID("some-dummy-value", "some.dummy.domain")
+    }
+
+    suspend fun withConversationDetailUpdate(conversationDetails: ConversationDetails) = apply {
+        coEvery { observeConversationDetails(any()) } returns conversationDetailsChannel.consumeAsFlow().map {
+            ObserveConversationDetailsUseCase.Result.Success(it)
+        }
+        conversationDetailsChannel.send(conversationDetails)
+        coEvery {
+            qualifiedIdMapper.fromStringToQualifiedID("id@domain")
+        } returns QualifiedID("id", "domain")
+    }
+
+    fun arrange() = this to viewModel
+}
